Fixed issues  according  to comments
