+++ Pipfile
@@ -14,4 +14,4 @@ pillow = "*"
 tensorflow = "*"
 keras ="*"
 [requires]
-python_version = "3.6"
+python_version = "3.7"
