+++ polyaxon/notebook.yml
@@ -9,8 +9,8 @@ environment:
       requests: 4
       limits: 4
     memory:
-      requests: 16000
-      limits: 20000
+      requests: 9000
+      limits: 10000
     gpu:
       requests: 1
       limits: 1
