+++ polyaxon/one_gpu_classification.yml
@@ -25,9 +25,6 @@ build:
     - pip install --no-cache-dir sklearn
 
 params:
-  input_path: /data/serendeepia/prolexitim_alexitimia/20190920_prolexytim_analytics
-  train_filename: train.csv
-  test_filename: test.csv
   models_path: /data/models/prolexitim_alexitimia
   tokenizer_file: tokenizer.pk
   embedding_path: /data/models/toxic_speech
@@ -35,7 +32,7 @@ params:
   embedding_matrix_file: embedding_matrix.pk
   weights_file: weights_base.best.hdf5
   max_features: 1000
-  max_length: 150
+  max_length: 284
   embed_size: 300
   epochs: 5
   patience: 300
@@ -45,4 +42,4 @@ params:
 
 run:
   cmd:
-    - python src/bilstm_train_classification.py --input_path {{ input_path }} --train_filename {{ train_filename }} --test_filename {{ test_filename }} --models_path {{ models_path }} --tokenizer_file {{ tokenizer_file }} --embedding_path {{ embedding_path }} --embedding_file {{ embedding_file }} --embedding_matrix_file {{ embedding_matrix_file }} --weights_file {{ weights_file }} --max_features {{ max_features }} --max_length {{ max_length }}  --embed_size {{ embed_size }}  --epochs {{ epochs }}  --batch_size {{ batch_size }}  --learning_rate {{ learning_rate }} --patience {{ patience }}
+    - python src/bilstm_train_classification.py  --models_path {{ models_path }} --tokenizer_file {{ tokenizer_file }} --embedding_path {{ embedding_path }} --embedding_file {{ embedding_file }} --embedding_matrix_file {{ embedding_matrix_file }} --weights_file {{ weights_file }} --max_features {{ max_features }} --max_length {{ max_length }}  --embed_size {{ embed_size }}  --epochs {{ epochs }}  --batch_size {{ batch_size }}  --learning_rate {{ learning_rate }} --patience {{ patience }}
