+++ polyaxon/one_gpu_regression.yml
@@ -25,9 +25,6 @@ build:
     - pip install --no-cache-dir sklearn
 
 params:
-  input_path: /data/serendeepia/prolexitim_alexitimia/20190920_prolexytim_analytics
-  train_filename: train.csv
-  test_filename: test.csv
   models_path: /data/models/prolexitim_alexitimia
   tokenizer_file: tokenizer.pk
   embedding_path: /data/models/toxic_speech
@@ -37,7 +34,7 @@ params:
   max_features: 1000
   max_length: 284
   embed_size: 300
-  epochs: 3
+  epochs: 1
   patience: 300
   batch_size: 128
   learning_rate: 0.001
@@ -46,4 +43,4 @@ params:
 
 run:
   cmd:
-    - python src/bilstm_train_regression.py --input_path {{ input_path }} --train_filename {{ train_filename }} --test_filename {{ test_filename }} --factor {{factor}} --models_path {{ models_path }} --tokenizer_file {{ tokenizer_file }} --embedding_path {{ embedding_path }} --embedding_file {{ embedding_file }} --embedding_matrix_file {{ embedding_matrix_file }} --weights_file {{ weights_file }} --max_features {{ max_features }} --max_length {{ max_length }}  --embed_size {{ embed_size }}  --epochs {{ epochs }}  --batch_size {{ batch_size }}  --learning_rate {{ learning_rate }} --patience {{ patience }}
+    - python src/bilstm_train_regression.py  --factor {{factor}} --models_path {{ models_path }} --tokenizer_file {{ tokenizer_file }} --embedding_path {{ embedding_path }} --embedding_file {{ embedding_file }} --embedding_matrix_file {{ embedding_matrix_file }} --weights_file {{ weights_file }} --max_features {{ max_features }} --max_length {{ max_length }}  --embed_size {{ embed_size }}  --epochs {{ epochs }}  --batch_size {{ batch_size }}  --learning_rate {{ learning_rate }} --patience {{ patience }}
