--- polyaxon/two_gpu.yml
@@ -1,48 +1,0 @@
----
-version: 1
-
-kind: experiment
-
-framework: tensorflow
-
-environment:
-  resources:
-    cpu:
-      requests: 8
-      limits: 10
-    memory:
-      requests: 16000
-      limits: 20000
-    gpu:
-      limits: 2
-
-build:
-  image: tensorflow/tensorflow:1.14.0-gpu-py3
-  build_steps:
-    - pip install --no-cache-dir -U polyaxon-cli
-    - pip install --no-cache-dir pandas
-    - pip install --no-cache-dir keras
-    - pip install --no-cache-dir sklearn
-
-params:
-  input_path: /data/serendeepia/prolexitim_alexitimia
-  train_filename: train.csv
-  test_filename: test.csv
-  models_path: /data/models/toxic_speech
-  tokenizer_file: tokenizer.pk
-  embedding_path: /data/models/toxic_speech
-  embedding_file: SBW-vectors-300-min5.txt
-  embedding_matrix_file: embedding_matrix.pk
-  weights_file: weights_base.best.hdf5
-  max_features: 100000
-  max_length: 150
-  embed_size: 300
-  epochs: 1
-  batch_size: 128
-  learning_rate: 0.001
-  patience: 100
-
-
-run:
-  cmd:
-    - python src/bilstm_train.py --input_path {{ input_path }} --train_filename {{ train_filename }} --test_filename {{ test_filename }} --models_path {{ models_path }} --tokenizer_file {{ tokenizer_file }} --embedding_path {{ embedding_path }} --embedding_file {{ embedding_file }} --embedding_matrix_file {{ embedding_matrix_file }} --weights_file {{ weights_file }} --max_features {{ max_features }} --max_length {{ max_length }}  --embed_size {{ embed_size }}  --epochs {{ epochs }}  --batch_size {{ batch_size }}  --learning_rate {{ learning_rate }} --patience {{ patience }}
